put some comments and made floats

+++ ruby_calc.rb
@@ -3,26 +3,27 @@ Calc_intro = "----Welcome to the Ruby Calculator----\nWhat is the first number"
 Which_operator = 'which operator would you like to use? (+ - * /)'
 Second_number = "What is the second number you would like to use?"
 @operator_array = ["+", "-", "*", "/"]
+
+#==========Calculator Answer Function==============
 def calc_answer
-  case @operator_in 
+  case @operator_in
     when "+"
       @answer = @first_num + @second_num
     when "-"
       @answer = @first_num - @second_num
-    when "*"    
+    when "*"
       @answer = @first_num * @second_num
     when "/"
-      @answer = @first_num / @second_num  
+      @answer = @first_num / @second_num
     else
       puts "Try Again"
-  end  
+  end
 end
 
-      
 #========MAIN PAGE===========
 def main_page
     puts Calc_intro
-    @first_num = gets.strip.to_i
+    @first_num = gets.strip.to_f
     operator
 end
 
@@ -45,12 +46,13 @@ end
 #==========SECOND NUMBER===========
 def second
   puts Second_number
-  @second_num = gets.strip.to_i
+  @second_num = gets.strip.to_f
   calculator
 end
 
+#============prints out calculator answer===============
 def calculator
-  puts "Calculating..." 
+  puts "Calculating..."
   calc_answer
   puts "the result of #{@first_num} #{@operator_in} #{@second_num} is #{@answer}"
 end
